Rename Helper to Logger and add tabulate print to Logger
